+++ constants.py
@@ -1,0 +1,1 @@
+DEFAULT_TIMEZONE = "Europe/Paris"
